+++ src/app/(crud)/UnitCRUD/components/ListUnit.tsx
@@ -1,8 +1,9 @@
 "use client";
 import React from "react";
-import { Card, Space, Table } from "antd";
+import { Card, Modal, Space, Table } from "antd";
 import type { ColumnsType } from "antd/es/table";
 import { IUnit } from "@/lib/interfaceBase";
+import { DeleteOutlined, EditOutlined } from "@ant-design/icons";
 
 interface IProps {
   onEdit: (unit: IUnit) => void;
@@ -15,7 +16,15 @@ const UnitController: React.FC<IProps> = ({ onEdit, onDelete, data }) => {
   };
 
   const handleDelete = (id: number) => {
-    onDelete(id);
+    Modal.confirm({
+      title: "Bạn có muốn xóa ?",
+      okText: "Yes",
+      okType: "danger",
+      width: "600px",
+      onOk: () => {
+        onDelete(id);
+      },
+    });
   };
   const columns: ColumnsType<IUnit> = [
     {
@@ -25,7 +34,7 @@ const UnitController: React.FC<IProps> = ({ onEdit, onDelete, data }) => {
       render: (text) => <a>{text}</a>,
     },
     {
-      title: "Name",
+      title: "Tên đơn vị",
       dataIndex: "name",
       key: "name",
     },
@@ -34,8 +43,8 @@ const UnitController: React.FC<IProps> = ({ onEdit, onDelete, data }) => {
       key: "action",
       render: (_, record) => (
         <Space size="middle">
-          <a onClick={() => handleEdit(record)}>Edit</a>
-          <a onClick={() => handleDelete(record.id)}>Delete</a>
+          <a onClick={() => handleEdit(record)}><EditOutlined/> Edit</a>
+          <a onClick={() => handleDelete(record.id)}><DeleteOutlined /> Delete</a>
         </Space>
       ),
     },
@@ -43,6 +52,7 @@ const UnitController: React.FC<IProps> = ({ onEdit, onDelete, data }) => {
   return (
     <Card>
       <Table
+        pagination={{defaultPageSize:5}}
         columns={columns}
         dataSource={data.map((unit) => ({ ...unit, key: unit.id }))}
       />
